requirements.txt
setuptools
mysql-connector-python
flask
flask_cors